Issue #2727 Cache mbeanFor lookup results
