+++ jetty-jmx/src/main/java/org/eclipse/jetty/jmx/MBeanContainer.java
@@ -19,19 +19,20 @@
 package org.eclipse.jetty.jmx;
 
 import java.io.IOException;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Objects;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Method;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import javax.management.InstanceNotFoundException;
-import javax.management.MBeanRegistrationException;
-import javax.management.MBeanServer;
-import javax.management.ObjectName;
+import javax.management.*;
+import javax.management.modelmbean.ModelMBean;
 
+import org.eclipse.jetty.util.Loader;
+import org.eclipse.jetty.util.annotation.ManagedAttribute;
 import org.eclipse.jetty.util.annotation.ManagedObject;
+import org.eclipse.jetty.util.annotation.ManagedOperation;
 import org.eclipse.jetty.util.component.Container;
 import org.eclipse.jetty.util.component.ContainerLifeCycle;
 import org.eclipse.jetty.util.component.Destroyable;
@@ -48,11 +49,115 @@ public class MBeanContainer implements Container.InheritedListener, Dumpable, De
     private final static Logger LOG = Log.getLogger(MBeanContainer.class.getName());
     private final static ConcurrentMap<String, AtomicInteger> __unique = new ConcurrentHashMap<>();
     private static final Container ROOT = new ContainerLifeCycle();
+    private static final Class<?>[] OBJ_ARG = new Class[]{Object.class};
 
     private final MBeanServer _mbeanServer;
+    private final ConcurrentMap<Class,Constructor<?>> _mbeanFor = new ConcurrentHashMap<>();
     private final ConcurrentMap<Object, Container> _beans = new ConcurrentHashMap<>();
     private final ConcurrentMap<Object, ObjectName> _mbeans = new ConcurrentHashMap<>();
     private String _domain = null;
+
+
+    /**
+     * <p>Creates an ObjectMBean for the given object.</p>
+     * <p>Attempts to create an ObjectMBean for the object by searching the package
+     * and class name space. For example an object of the type:</p>
+     * <pre>
+     * class com.acme.MyClass extends com.acme.util.BaseClass implements com.acme.Iface
+     * </pre>
+     * <p>then this method would look for the following classes:</p>
+     * <ul>
+     * <li>com.acme.jmx.MyClassMBean</li>
+     * <li>com.acme.util.jmx.BaseClassMBean</li>
+     * <li>org.eclipse.jetty.jmx.ObjectMBean</li>
+     * </ul>
+     *
+     * @param o The object
+     * @return A new instance of an MBean for the object or null.
+     */
+    public Object mbeanFor(Object o)
+    {
+        if (o==null)
+            return null;
+        
+        Class<?> oClass = o.getClass();
+
+        Constructor<?> constructor = _mbeanFor.computeIfAbsent(oClass,c->
+        {
+            while (true)
+            {
+                try
+                {
+                    String pName = c.getPackage().getName();
+                    String cName = c.getName().substring(pName.length() + 1);
+                    String mName = pName + ".jmx." + cName + "MBean";
+
+                    Class<?> mClass;
+                    try
+                    {
+                        // Look for an MBean class from the same loader that loaded the original class
+                        mClass = (Object.class.equals(c))?c = ObjectMBean.class:Loader.loadClass(c, mName);
+                    }
+                    catch (ClassNotFoundException e)
+                    {
+                        // Not found, so if not the same as the thread context loader, try that.
+                        if (Thread.currentThread().getContextClassLoader() == c.getClassLoader())
+                            throw e;
+                        LOG.ignore(e);
+                        mClass = Loader.loadClass(c, mName);
+                    }
+
+                    if (LOG.isDebugEnabled())
+                        LOG.debug("ObjectMBean: mbeanFor {} mClass={}", o, mClass);
+
+                    return (ModelMBean.class.isAssignableFrom(mClass))
+                        ?mClass.getDeclaredConstructor()
+                        :mClass.getConstructor(OBJ_ARG);
+                }
+                catch (ClassNotFoundException | NoSuchMethodException e)
+                {
+                    LOG.debug(e.toString());
+                    LOG.ignore(e);
+                }
+
+                c = c.getSuperclass();
+                if (c==null)
+                    return null;
+            }
+        });
+
+        if (constructor==null)
+            return null;
+
+        try
+        {
+            Object mbean;
+            if (constructor.getParameterCount()==0)
+            {
+                mbean = constructor.newInstance();
+                ((ModelMBean)mbean).setManagedResource(o, "objectReference");
+            }
+            else
+            {
+                mbean = constructor.newInstance(o);
+            }
+
+            if (mbean instanceof ObjectMBean)
+                ((ObjectMBean)mbean).setMBeanContainer(this);
+
+            if (LOG.isDebugEnabled())
+                LOG.debug("mbeanFor {} is {}", o, mbean);
+
+            return mbean;
+        }
+        catch(Throwable t)
+        {
+            LOG.warn(t);
+        }
+
+        return null;
+    }
+
 
     /**
      * Lookup an object name by instance
@@ -154,14 +259,13 @@ public class MBeanContainer implements Container.InheritedListener, Dumpable, De
         try
         {
             // Create an MBean for the object.
-            Object mbean = ObjectMBean.mbeanFor(obj);
+            Object mbean = mbeanFor(obj);
             if (mbean == null)
                 return;
 
             ObjectName objectName = null;
             if (mbean instanceof ObjectMBean)
             {
-                ((ObjectMBean)mbean).setMBeanContainer(this);
                 objectName = ((ObjectMBean)mbean).getObjectName();
             }
 
@@ -252,6 +356,111 @@ public class MBeanContainer implements Container.InheritedListener, Dumpable, De
                 .replace(',', '_')
                 .replace(' ', '_');
     }
+
+
+    MBeanInfo getMBeanInfo(ObjectMBean bean)
+    {
+        String desc = null;
+        List<MBeanAttributeInfo> attributes = new ArrayList<>();
+        List<MBeanOperationInfo> operations = new ArrayList<>();
+
+        // Find list of classes that can influence the mbean
+        Object managed = bean.getManagedObject();
+        Class<?> o_class = managed.getClass();
+        List<Class<?>> influences = new ArrayList<>();
+        influences.add(bean.getClass()); // always add MBean itself
+        influences = findInfluences(influences, managed.getClass());
+
+        if (LOG.isDebugEnabled())
+            LOG.debug("Influence Count: {}", influences.size());
+
+        // Process Type Annotations
+        ManagedObject primary = o_class.getAnnotation(ManagedObject.class);
+
+        if (primary != null)
+        {
+            desc = primary.value();
+        }
+        else
+        {
+            if (LOG.isDebugEnabled())
+                LOG.debug("No @ManagedObject declared on {}", managed.getClass());
+        }
+
+        // For each influence
+        for (Class<?> oClass : influences)
+        {
+            ManagedObject typeAnnotation = oClass.getAnnotation(ManagedObject.class);
+
+            if (LOG.isDebugEnabled())
+                LOG.debug("Influenced by: " + oClass.getCanonicalName());
+
+            if (typeAnnotation == null)
+            {
+                if (LOG.isDebugEnabled())
+                    LOG.debug("Annotations not found for: {}", oClass.getCanonicalName());
+                continue;
+            }
+
+            // Process Method Annotations
+
+            for (Method method : oClass.getDeclaredMethods())
+            {
+                ManagedAttribute methodAttributeAnnotation = method.getAnnotation(ManagedAttribute.class);
+
+                if (methodAttributeAnnotation != null)
+                {
+                    if (LOG.isDebugEnabled())
+                        LOG.debug("Attribute Annotation found for: {}", method.getName());
+                    MBeanAttributeInfo mai = bean.defineAttribute(method, methodAttributeAnnotation);
+                    if (mai != null)
+                        attributes.add(mai);
+                }
+
+                ManagedOperation methodOperationAnnotation = method.getAnnotation(ManagedOperation.class);
+
+                if (methodOperationAnnotation != null)
+                {
+                    if (LOG.isDebugEnabled())
+                        LOG.debug("Method Annotation found for: {}", method.getName());
+                    MBeanOperationInfo oi = bean.defineOperation(method, methodOperationAnnotation);
+                    if (oi != null)
+                        operations.add(oi);
+                }
+            }
+        }
+
+        return new MBeanInfo(o_class.getName(),
+            desc,
+            attributes.toArray(new MBeanAttributeInfo[attributes.size()]),
+            new MBeanConstructorInfo[0],
+            operations.toArray(new MBeanOperationInfo[operations.size()]),
+            new MBeanNotificationInfo[0]);
+    }
+
+
+    private static List<Class<?>> findInfluences(List<Class<?>> influences, Class<?> aClass)
+    {
+        if (aClass != null)
+        {
+            if (!influences.contains(aClass))
+            {
+                // This class is a new influence
+                influences.add(aClass);
+            }
+
+            // So are the super classes
+            influences = findInfluences(influences, aClass.getSuperclass());
+
+            // So are the interfaces
+            Class<?>[] ifs = aClass.getInterfaces();
+            for (int i = 0; ifs != null && i < ifs.length; i++)
+                influences = findInfluences(influences, ifs[i]);
+        }
+        return influences;
+    }
+
+
 
     @Override
     public void dump(Appendable out, String indent) throws IOException
