+++ jetty-jmx/src/test/java/org/eclipse/jetty/jmx/ObjectMBeanUtilTest.java
@@ -113,7 +113,7 @@ public class ObjectMBeanUtilTest
     public void testMbeanForNullCheck()
     {
         // when
-        mBean = ObjectMBean.mbeanFor(null);
+        mBean = container.mbeanFor(null);
 
         // then
         assertNull(mBean, "As we are passing null value the output should be null");
@@ -351,7 +351,7 @@ public class ObjectMBeanUtilTest
 
     private void setMBeanInfoForInvoke()
     {
-        objectMBean = (ObjectMBean)ObjectMBean.mbeanFor(derivedExtended);
+        objectMBean = (ObjectMBean)container.mbeanFor(derivedExtended);
         container.beanAdded(null,derivedExtended);
         objectMBean.getMBeanInfo();
     }
