+++ jetty-jmx/src/test/resources/jetty-logging.properties
@@ -1,2 +1,2 @@
 org.eclipse.jetty.util.log.class=org.eclipse.jetty.util.log.StdErrLog
-org.eclipse.jetty.jmx.LEVEL=INFO
+org.eclipse.jetty.jmx.LEVEL=ALL
